pyproject.toml
[tool.poetry]
name = "nijisanji-voice-organize"
version = "0.1.0"
description = ""
authors = ["iron316 <shunta316iron@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pysen = {version = "0.9.1", extras = ["lint"]}
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
